Fix compiler erorrs

+++ common/include/readprofile.h
@@ -85,7 +85,7 @@ const QMap<QString, WORD> keyMap = {
 //     return map;
 // }();
 #elif defined(__APPLE__)
-const QMap<QString, WORD> keyMap = {};
+const QMap<QString, int> keyMap = {};
 #elif defined(__linux__)
 const QMap<QString, WORD> keyMap = {};
 #else
